加入购物车接口对接

+++ app/src/main/java/com/wzy/mhealth/model/Shop.java
@@ -17,7 +17,7 @@ public class Shop {
 
     /**
      * showImage : http://117.34.105.29:8818/mhealth/upload/files/20161122135940xCwNnP66.png
-     * productData : [{"productImageBig":"http://117.34.105.29:8818/mhealth/upload/files/20161121104516jH15N1YD.png","productNewPrice":258,"productOldPrice":298,"productImageSmall":"http://117.34.105.29:8818/mhealth/upload/files/20161121104515iMnDUMs8.png","productName":"有机青稞米2.5kg礼盒装","productId":"f5a2e99d5884b056015884c5f2c80003"},{"productImageBig":"http://117.34.105.29:8818/mhealth/upload/files/20161121104516jH15N1YD.png","productNewPrice":258,"productOldPrice":298,"productImageSmall":"http://117.34.105.29:8818/mhealth/upload/files/20161121104515iMnDUMs8.png","productName":"有机青稞米2.5kg礼盒装","productId":"f5a2e99d5884b056015884c5f2c80003"},{"productImageBig":"http://117.34.105.29:8818/mhealth/upload/files/20161121104516jH15N1YD.png","productNewPrice":258,"productOldPrice":298,"productImageSmall":"http://117.34.105.29:8818/mhealth/upload/files/20161121104515iMnDUMs8.png","productName":"有机青稞米2.5kg礼盒装","productId":"f5a2e99d5884b056015884c5f2c80003"},{"productImageBig":"http://117.34.105.29:8818/mhealth/upload/files/20161121104516jH15N1YD.png","productNewPrice":258,"productOldPrice":298,"productImageSmall":"http://117.34.105.29:8818/mhealth/upload/files/20161121104515iMnDUMs8.png","productName":"有机青稞米2.5kg礼盒装","productId":"f5a2e99d5884b056015884c5f2c80003"}]
+     * productData : [{"productImageBig":"http://117.34.105.29:8818/mhealth/upload/files/20161121102139jprZEZCB.png","productNewPrice":10.8,"productOldPrice":12.8,"productImageSmall":"http://117.34.105.29:8818/mhealth/upload/files/20161121102139yPcnfkSW.png","productName":"有机青稞米100g","productId":"f5a2e99d5884b056015884b056520000"},{"productImageBig":"http://117.34.105.29:8818/mhealth/upload/files/201611211030474Zujn8h6.png","productNewPrice":49.8,"productOldPrice":50.8,"productImageSmall":"http://117.34.105.29:8818/mhealth/upload/files/20161121103048rLwq5yDr.png","productName":"有机青稞米500g","productId":"f5a2e99d5884b056015884b8ac890001"},{"productImageBig":"http://117.34.105.29:8818/mhealth/upload/files/20161121103827ghk0Tp9g.png","productNewPrice":228,"productOldPrice":268,"productImageSmall":"http://117.34.105.29:8818/mhealth/upload/files/20161121103826mS7CmMQL.png","productName":"有机青稞米2.5kg家庭版","productId":"f5a2e99d5884b056015884bfb8150002"},{"productImageBig":"http://117.34.105.29:8818/mhealth/upload/files/20161121104516jH15N1YD.png","productNewPrice":258,"productOldPrice":298,"productImageSmall":"http://117.34.105.29:8818/mhealth/upload/files/20161121104515iMnDUMs8.png","productName":"有机青稞米2.5kg礼盒装","productId":"f5a2e99d5884b056015884c5f2c80003"}]
      */
 
     private List<DataEntity> data;
@@ -33,12 +33,12 @@ public class Shop {
     public static class DataEntity {
         private String showImage;
         /**
-         * productImageBig : http://117.34.105.29:8818/mhealth/upload/files/20161121104516jH15N1YD.png
-         * productNewPrice : 258
-         * productOldPrice : 298
-         * productImageSmall : http://117.34.105.29:8818/mhealth/upload/files/20161121104515iMnDUMs8.png
-         * productName : 有机青稞米2.5kg礼盒装
-         * productId : f5a2e99d5884b056015884c5f2c80003
+         * productImageBig : http://117.34.105.29:8818/mhealth/upload/files/20161121102139jprZEZCB.png
+         * productNewPrice : 10.8
+         * productOldPrice : 12.8
+         * productImageSmall : http://117.34.105.29:8818/mhealth/upload/files/20161121102139yPcnfkSW.png
+         * productName : 有机青稞米100g
+         * productId : f5a2e99d5884b056015884b056520000
          */
 
         private List<ProductDataEntity> productData;
@@ -61,8 +61,8 @@ public class Shop {
 
         public static class ProductDataEntity {
             private String productImageBig;
-            private int productNewPrice;
-            private int productOldPrice;
+            private double productNewPrice;
+            private double productOldPrice;
             private String productImageSmall;
             private String productName;
             private String productId;
@@ -71,11 +71,11 @@ public class Shop {
                 this.productImageBig = productImageBig;
             }
 
-            public void setProductNewPrice(int productNewPrice) {
+            public void setProductNewPrice(double productNewPrice) {
                 this.productNewPrice = productNewPrice;
             }
 
-            public void setProductOldPrice(int productOldPrice) {
+            public void setProductOldPrice(double productOldPrice) {
                 this.productOldPrice = productOldPrice;
             }
 
@@ -95,11 +95,11 @@ public class Shop {
                 return productImageBig;
             }
 
-            public int getProductNewPrice() {
+            public double getProductNewPrice() {
                 return productNewPrice;
             }
 
-            public int getProductOldPrice() {
+            public double getProductOldPrice() {
                 return productOldPrice;
             }
